single-gpu-transformer/requirements.txt
PYYAML
optax
flax
ml_collections
